Move initial RoomService to separated folder
